% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-model-output.R
\name{get_model_output}
\alias{get_model_output}
\title{Specify a fit model to submit to the validation server}
\usage{
get_model_output(fit = NULL, model_name = NULL, validate = TRUE)
}
\arguments{
\item{fit}{A model object. Currently only \code{lm} and \code{glm} are supported.}

\item{model_name}{Arbitrary name for the model}
}
\value{
A data frame formatted for the validation server
}
\description{
Specify a fit model to submit to the validation server
}
