% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table-output.R
\name{get_table_output}
\alias{get_table_output}
\title{Specify a tabular analysis to submit to the validation server}
\usage{
get_table_output(
  data = NULL,
  table_name = NULL,
  stat = NULL,
  var = NULL,
  by = NULL,
  na.rm = FALSE,
  validate = TRUE
)
}
\arguments{
\item{data}{Data frame object}

\item{table_name}{Arbitrary name for the analysis}

\item{stat}{Summary statistic(s) to calculate (e.g. mean, median, sum, sd, var, min, max, n, n_distinct)}

\item{var}{Variable(s) to summarize}

\item{by}{Optional variable(s) to group by}

\item{na.rm}{Optional NA behavior (defaults to FALSE)}
}
\value{
A data frame formatted for the validation server
}
\description{
Specify a tabular analysis to submit to the validation server
}
