% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-output.R
\name{submit_output}
\alias{submit_output}
\title{Combine table and model analyses to submit to the validation server}
\usage{
submit_output(...)
}
\arguments{
\item{...}{Data frames returned from the get_table_output() and/or get_model_output() functions}
}
\value{
A combined data frame formatted for the validation server
}
\description{
Combine table and model analyses to submit to the validation server
}
