% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-model-output.R
\name{validate_model_args}
\alias{validate_model_args}
\title{Validate get_model_output() arguments with basic error handling}
\usage{
validate_model_args(fit, model_name)
}
\description{
Validate get_model_output() arguments with basic error handling
}
