% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table-output.R
\name{validate_table_args}
\alias{validate_table_args}
\title{Validate get_table_output() arguments with basic error handling}
\usage{
validate_table_args(data, table_name, stat, var, by, stat_fn)
}
\description{
Validate get_table_output() arguments with basic error handling
}
